add tag for build
