+++ common/net/dnsconf.go
@@ -1,0 +1,7 @@
+// +build !android
+
+package net
+
+func SetDefaultResolver(ns []byte) string {
+	return ""
+}
